@@ -1,34 +1,37 @@
 {
   "name": "typograf-constructor",
-  "description": "Fastest way to create your rule for typograf ",
+  "description": "Fastest way to create your rules for Typograf",
   "version": "1.0.0",
   "author": {
     "name": "Denis Seleznev",
     "email": "hcodes@yandex.ru",
     "url": "https://github.com/typograf/typograf-constructor"
   },
-  "homepage": "https://github.com/typograf/constructor",
-  "licenses": [
-    {
-      "type": "MIT",
-      "url": "https://github.com/typograf/constructor/blob/master/LICENSE"
-    }
-  ],
+  "homepage": "https://typograf.github.io/constructor/",
+  "licenses": "MIT",
   "repository": {
     "type": "git",
     "url": "git://github.com/typograf/typograf-constructor.git"
   },
   "engines": {
-    "node": ">= 0.10"
+    "node": ">= 4.0"
   },
   "devDependencies": {
     "eslint": "^0.21.2",
     "eslint-plugin-react": "^2.3.0",
-    "react-tools": "^0.13.3"
+    "react-scripts": "0.6.0"
+  },
+  "dependencies": {
+    "typograf": "*",
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2"
   },
   "scripts": {
-    "test": "npm run-script eslint",
+    "build": "rd \"static\" /S /Q && del \"index.html\" /Q && react-scripts build && xcopy \"build\" . /S /I",
     "eslint": "./node_modules/.bin/eslint .",
-    "build": "./node_modules/.bin/jsx ./src ./build"
-  }
+    "test": "npm run-script eslint && react-scripts test --env=jsdom",
+    "start": "react-scripts start",
+    "eject": "react-scripts eject"
+  },
+  "private": true
 }
